refactor(log): move Init function to internal package
Relocate the logger initialization logic from the public log package
to internal/log to align with the architecture where initialization
functions reside in internal packages. The public log package now
contains only the context-aware logger utilities. Add comprehensive
tests for parseLevel and Init functions.
